[#10] Make it impossible to re-engage a vaultdoor
This is specifically regarding the situation where you are cracking a
vault door, and then run into the door again.

This did not have any practical effect yet.
However, it would allow giving the enemies extra turns, which was never
the intention (yet is still possible when cracking a vault, as you can
just re-guess the wrong digit).
I encountered it a bit earlier, though, as there was a situation that
would become easier to handle if I didn't have to take this strange move
into account.